third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-getSVGDocument-iframe.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
function runTest() {
    test(() => {
      assert_equals(document.getElementById("svgobject").getSVGDocument(), null);
    }, "getSVGDocument() returns null for cross-origin documents");
}
</script>
</head>
<body onload="runTest()">
<iframe id="svgobject" src="http://localhost:8080/security/resources/empty-svg.php"
        width="400" height="300"></iframe> 
</body>
</html>
